FREEMARKER-188 [freemarker-generator] Support an output "generation" mode
