+++ freemarker-generator-base/src/main/java/org/apache/freemarker/generator/base/FreeMarkerConstants.java
@@ -93,6 +93,15 @@ public class FreeMarkerConstants {
         public static final String FREEMARKER_USER_PARAMETERS = "freemarker.user.parameters";
     }
 
+    public static class Mode {
+
+        private Mode() {
+        }
+
+        public static final String AGGREGATE = "aggregate";
+        public static final String GENERATE = "generate";
+    }
+
     public static class SystemProperties {
 
         private SystemProperties() {
