+++ freemarker-generator-cli/src/app/templates/freemarker-generator/info.ftl
@@ -25,10 +25,16 @@ Timestamp              : ${.now}
 Output encoding        : ${.output_encoding}
 Output format          : ${.output_format}
 
+FreeMarker Command-line Parameters
+------------------------------------------------------------------------------
+<#list tools.system.getCommandLineArgs() as arg>
+    [#${arg?counter}] ${arg}
+</#list>
+
 FreeMarker Generator Template Loader Directories
 ------------------------------------------------------------------------------
 <#list tools.system.getTemplateDirectories() as directory>
-[#${directory?counter}] ${directory}
+    [#${directory?counter}] ${directory}
 </#list>
 
 FreeMarker Generator Data Model
